stats widget done, edit dialog boxes WIP
