+++ lib/widgets/btm_nav/home_widget.dart
@@ -40,6 +40,7 @@ class HomeWidget extends StatelessWidget {
               AnnouncementBanner(),
               SmallShortcutsRowWidget(),
               // SmallShortcutsRowWidget()
+              // StreakWidget(),
               StatsWidget()
             ],
           ),
